# cmd/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Meetin APP API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/meeting/CustomerServiceHistory/multiple": {
            "put": {
                "description": "Update Multiple in UpdateCustomerServiceHistory",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Update Multiple in CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistorysUpdateMultiple",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.CustomerServiceHistoryEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create Multiple in CreateCustomerServiceHistory",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Create Multiple in CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistoryCreateMultiple",
                "parameters": [
                    {
                        "type": "string",
                        "default": "[[authToken]]",
                        "description": "Current Session Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "[[companyID]]",
                        "description": "Current CompanyID",
                        "name": "x-company",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.CustomerServiceHistoryFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/CustomerServiceHistory/pagination": {
            "get": {
                "description": "Pagination GetCustomerServiceHistoryWithPagination",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Pagination CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistoryPagination",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/CustomerServiceHistory/search": {
            "get": {
                "description": "Get GetCustomerServiceHistory",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Show CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistorySearch",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/CustomerServiceHistory/single": {
            "post": {
                "description": "Create in CreateCustomerServiceHistory",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Create in CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistoryCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.CustomerServiceHistoryFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/meeting/CustomerServiceHistory/with-id/:id": {
            "get": {
                "description": "Get GetCustomerServiceHistoryById",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Get CustomerServiceHistory by Unique IDs",
                "operationId": "ApiV1MeetingAppCustomerServiceHistoryWithID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update with id in UpdateCustomerServiceHistory",
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Update with id in CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistoryUpdateWithID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "description": "update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_structures.CustomerServiceHistoryEdit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_structures.CustomerServiceHistory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete with path params in DeleteCustomerServiceHistory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meeting CustomerServiceHistory"
                ],
                "summary": "Delete with path params in CustomerServiceHistory",
                "operationId": "ApiV1MeetingAppCustomerServiceHistoryDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id uuid",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_structures.CustomerServiceHistory": {
            "type": "object",
            "properties": {
                "appointmentId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "customerId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "notes": {
                    "type": "string",
                    "example": "Quisquam voluptatem."
                },
                "serviceDate": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "serviceId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
        "api_structures.CustomerServiceHistoryEdit": {
            "type": "object",
            "properties": {
                "appointmentId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "customerId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "notes": {
                    "type": "string",
                    "example": "Quisquam voluptatem."
                },
                "serviceDate": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "serviceId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
        "api_structures.CustomerServiceHistoryFilter": {
            "type": "object",
            "properties": {
                "appointmentId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "createdAt": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "customerId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "deletedAt": {
                    "type": "string",
                    "example": "1989-07-30T10:05:53Z"
                },
                "id": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "notes": {
                    "type": "string",
                    "example": "Quisquam voluptatem."
                },
                "serviceDate": {
                    "type": "string",
                    "example": "1960-11-10T04:48:09Z"
                },
                "serviceId": {
                    "type": "string",
                    "example": "e4a3b2c1-3b2c-4a3b-a2c1-1b2c3a4b5c6d"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "1986-08-02T22:27:19Z"
                }
            }
        },
...
    }
}